+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 All notable changes to this project will be documented in this file.
 
 ## [1.1.5] - 2021-05-01
--   types fix.
--   delete and restore fix.
+-   fixed types.
+-   fixed delete and restore.
 
 ## [1.1.3] - 2021-04-30
 -   fixed bad nesting handling.