+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file.
 
-## [1.1.5] - 2021-05-01
+## [1.1.6] - 2021-05-01
 -   fixed types.
 -   fixed delete and restore.
 -   added array support.